Legacy hack quick fix
This is an unfortunate and innocent consequence of the framework update, which was intended to be as unobtrusive as possible and simply update management of hacks behind the scenes. Part of the update was to move the enabled state away from the config, so the enabled value is only used to determine whether the hack *should* be enabled, not whether it *is* enabled. And so since TimingsHacks is disabled by default (I assume) and is riddled with `if (config.isEnabled()) return;` means that it's enabled but nothing happens. This introduces a quick fix, adding legacy support for `this.config.isEnabled()` state tracking though making it explicitly deprecated. That being said, most of the hacks are in dire need of an update and clean up.